Rename parser.ts to parseNode.ts
